requirements.txt
hidapi
psutil